server rebuild/recompilation now optional in runserver.sh script

+++ webserver/runserver.sh
@@ -5,8 +5,11 @@ set -x
 
 pkgname=bibleServer
 
-rm -rfv $pkgname
-javac src/* -d .
+if [ "$1" = "--rebuild" ] || [ "$1" = "-r" ] ; then
+    rm -rfv $pkgname
+    javac src/* -d .
+fi
+
 mkdir -p "./server-logs"
 java $pkgname.BibleServer -p 8080 -r "./public" \
     | tee "./server-logs/server_log_started_on_$( date | tr ' ' '-' ).log"
